Fixed API script, hopefully once and for all. Also added a second key to use.

+++ Server/sql/db_create_tables.sql
@@ -1,6 +1,6 @@
 --this file creates all tables with their appropriate types and constraints
 create table users (
-    id int not null auto_increment,
+    id bigint not null auto_increment,
     username varchar(50) not null unique,
     email varchar(100) not null unique,
     password varchar(150) not null,
@@ -10,7 +10,7 @@ create table users (
 );
 
 create table items (
-    id int not null auto_increment,
+    id bigint not null auto_increment,
     type varchar(15) not null,
     title varchar(100) not null,
     description text not null,
@@ -19,14 +19,14 @@ create table items (
 );
 
 create table genres (
-    id int not null auto_increment,
+    id bigint not null auto_increment,
     name varchar(15) unique,
     primary key(id)
 );
 
 create table comments (
-    id int not null auto_increment,
-    user_id int not null,
+    id bigint not null auto_increment,
+    user_id bigint not null,
     create_date datetime not null,
     content text not null,
     primary key(id),
@@ -34,18 +34,18 @@ create table comments (
 );
 
 create table comment_parents (
-    id int not null auto_increment,
-    parent_id int not null,
-    child_id int not null,
+    id bigint not null auto_increment,
+    parent_id bigint not null,
+    child_id bigint not null,
     primary key(id),
     foreign key(parent_id) references comments(id),
     foreign key(child_id) references comments(id)
 );
 
 create table comment_ratings (
-    id int not null auto_increment,
-    user_id int not null,
-    comment_id int not null,
+    id bigint not null auto_increment,
+    user_id bigint not null,
+    comment_id bigint not null,
     rating boolean not null,
     primary key(id),
     foreign key(user_id) references users(id),
@@ -53,8 +53,8 @@ create table comment_ratings (
 );
 
 create table libraries (
-    id int not null auto_increment,
-    user_id int not null,
+    id bigint not null auto_increment,
+    user_id bigint not null,
     library_name varchar(30),
     visible boolean default true not null,
     primary key(id),
@@ -62,17 +62,17 @@ create table libraries (
 );
 
 create table library_items (
-    id int not null auto_increment,
-    library_id int not null,
-    item_id int not null,
+    id bigint not null auto_increment,
+    library_id bigint not null,
+    item_id bigint not null,
     primary key(id),
     foreign key(library_id) references libraries(id),
     foreign key(item_id) references items(id)
 );
 
 create table item_statistics (
-    id int not null auto_increment,
-    item_id int not null,
+    id bigint not null auto_increment,
+    item_id bigint not null,
     last_viewed datetime not null,
     views int not null,
     primary key(id),
@@ -80,26 +80,26 @@ create table item_statistics (
 );
 
 create table item_images (
-    id int not null auto_increment,
-    item_id int not null,
+    id bigint not null auto_increment,
+    item_id bigint not null,
     image_url varchar(200) not null,
     primary key(id),
     foreign key(item_id) references items(id)
 );
 
 create table item_genres (
-    id int not null auto_increment,
-    item_id int not null,
-    genre_id int not null,
+    id bigint not null auto_increment,
+    item_id bigint not null,
+    genre_id bigint not null,
     primary key(id),
     foreign key(item_id) references items(id),
     foreign key(genre_id) references genres(id)
 );
 
 create table item_reviews (
-    id int not null auto_increment,
-    item_id int not null,
-    user_id int not null,
+    id bigint not null auto_increment,
+    item_id bigint not null,
+    user_id bigint not null,
     review_value int,
     primary key(id),
     foreign key(item_id) references items(id),
@@ -107,26 +107,26 @@ create table item_reviews (
 );
 
 create table item_comments (
-    id int not null auto_increment,
-    item_id int not null,
-    comment_id int not null,
+    id bigint not null auto_increment,
+    item_id bigint not null,
+    comment_id bigint not null,
     primary key(id),
     foreign key(item_id) references items(id),
     foreign key(comment_id) references comments(id)
 );
 
 create table user_ignored_items (
-    id int not null auto_increment,
-    user_id int not null,
-    item_id int not null,
+    id bigint not null auto_increment,
+    user_id bigint not null,
+    item_id bigint not null,
     primary key(id),
     foreign key(item_id) references items(id),
     foreign key(user_id) references users(id)
 );
 
 create table user_sessions (
-    id int not null auto_increment,
-    user_id int not null,
+    id bigint not null auto_increment,
+    user_id bigint not null,
     session_id varchar(100),
     session_exp datetime null,
     primary key(id),
@@ -134,28 +134,28 @@ create table user_sessions (
 );
 
 create table user_friends (
-    id int not null auto_increment,
-    user_id int not null,
-    friend_id int not null,
+    id bigint not null auto_increment,
+    user_id bigint not null,
+    friend_id bigint not null,
     primary key(id),
     foreign key(user_id) references users(id),
     foreign key(user_id) references users(id)
 );
 
 create table user_parental_controls (
-    id int not null auto_increment,
-    user_id int not null,
+    id bigint not null auto_increment,
+    user_id bigint not null,
     rating varchar(10),
-    genre_id int not null,
+    genre_id bigint not null,
     primary key(id),
     foreign key(user_id) references users(id),
     foreign key(genre_id) references genres(id)
 );
 
 create table item_children (
-    id int not null auto_increment,
-    parent_item_id int not null,
-    child_item_id int not null,
+    id bigint not null auto_increment,
+    parent_item_id bigint not null,
+    child_item_id bigint not null,
     primary key(id),
     foreign key(parent_item_id) references items(id),
     foreign key(child_item_id) references items(id)
